migrate platform-dependent code to separate modules
