+++ src/uu/split/src/platform/mod.rs
@@ -1,0 +1,11 @@
+#[cfg(unix)]
+pub use self::unix::instantiate_current_writer;
+
+#[cfg(windows)]
+pub use self::windows::instantiate_current_writer;
+
+#[cfg(unix)]
+mod unix;
+
+#[cfg(windows)]
+mod windows;
