+++ src/uu/split/src/split.rs
@@ -10,16 +10,13 @@
 #[macro_use]
 extern crate uucore;
 
+mod platform;
+
 use std::char;
-#[cfg(unix)]
 use std::env;
-use std::fs::{File, OpenOptions};
-#[cfg(unix)]
-use std::io::Result;
+use std::fs::File;
 use std::io::{stdin, stdout, BufRead, BufReader, BufWriter, Read, Write};
 use std::path::Path;
-#[cfg(unix)]
-use std::process::{Child, Command, Stdio};
 
 static NAME: &str = "split";
 static VERSION: &str = env!("CARGO_PKG_VERSION");
@@ -311,142 +308,6 @@ fn num_prefix(i: usize, width: usize) -> String {
     c
 }
 
-#[cfg(unix)]
-/// A writer that writes to a shell_process' stdin
-///
-/// We use a shell process (not directy calling a sub-process) so we can forward the name of the
-/// corresponding output file (xaa, xab, xac… ). This is the way it was implemented in GNU split.
-struct FilterWriter {
-    /// Running shell process
-    shell_process: Child,
-}
-
-#[cfg(unix)]
-impl Write for FilterWriter {
-    fn write(&mut self, buf: &[u8]) -> Result<usize> {
-        self.shell_process
-            .stdin
-            .as_mut()
-            .expect("failed to get shell stdin")
-            .write(buf)
-    }
-    fn flush(&mut self) -> Result<()> {
-        self.shell_process
-            .stdin
-            .as_mut()
-            .expect("failed to get shell stdin")
-            .flush()
-    }
-}
-
-/// Have an environment variable set at a value during this lifetime
-#[cfg(unix)]
-struct WithEnvVarSet {
-    /// Env var key
-    _previous_var_key: String,
-    /// Previous value set to this key
-    _previous_var_value: std::result::Result<String, env::VarError>,
-}
-#[cfg(unix)]
-impl WithEnvVarSet {
-    /// Save previous value assigned to key, set key=value
-    fn new(key: &str, value: &str) -> WithEnvVarSet {
-        let previous_env_value = env::var(key);
-        env::set_var(key, value);
-        WithEnvVarSet {
-            _previous_var_key: String::from(key),
-            _previous_var_value: previous_env_value,
-        }
-    }
-}
-#[cfg(unix)]
-impl Drop for WithEnvVarSet {
-    /// Restore previous value now that this is being dropped by context
-    fn drop(&mut self) {
-        if let Ok(ref prev_value) = self._previous_var_value {
-            env::set_var(&self._previous_var_key, &prev_value);
-        } else {
-            env::remove_var(&self._previous_var_key)
-        }
-    }
-}
-
-#[cfg(unix)]
-impl FilterWriter {
-    /// Create a new filter running a command with $FILE pointing at the output name
-    ///
-    /// #Arguments
-    ///
-    /// * `command` - The shell command to execute
-    /// * `filepath` - Path of the output file (forwarded to command as $FILE)
-    fn new(command: &str, filepath: &str) -> FilterWriter {
-        // set $FILE, save previous value (if there was one)
-        let _with_env_var_set = WithEnvVarSet::new("FILE", &filepath);
-
-        let shell_process = Command::new(env::var("SHELL").unwrap_or("/bin/sh".to_owned()))
-            .arg("-c")
-            .arg(command)
-            .stdin(Stdio::piped())
-            .spawn()
-            .expect("Couldn't spawn filter command");
-
-        FilterWriter {
-            shell_process: shell_process,
-        }
-    }
-}
-
-#[cfg(unix)]
-impl Drop for FilterWriter {
-    /// flush stdin, close it and wait on `shell_process` before dropping self
-    fn drop(&mut self) {
-        {
-            // close stdin by dropping it
-            let _stdin = self.shell_process.stdin.as_mut();
-        }
-        let exit_status = self
-            .shell_process
-            .wait()
-            .expect("Couldn't wait for child process");
-        if let Some(return_code) = exit_status.code() {
-            if return_code != 0 {
-                crash!(1, "Shell process returned {}", return_code);
-            }
-        } else {
-            crash!(1, "Shell process terminated by signal")
-        }
-    }
-}
-
-fn get_file_writer(filename: &str) -> BufWriter<Box<dyn Write>> {
-    BufWriter::new(Box::new(
-        // write to the next file
-        OpenOptions::new()
-            .write(true)
-            .create(true)
-            .open(Path::new(&filename))
-            .unwrap(),
-    ) as Box<dyn Write>)
-}
-
-#[cfg(windows)]
-fn instantiate_current_writer(_settings: &Settings, filename: &str) -> BufWriter<Box<dyn Write>> {
-    get_file_writer(filename)
-}
-
-/// Instantiate either a file writer or a "write to shell process's stdin" writer
-#[cfg(unix)]
-fn instantiate_current_writer(settings: &Settings, filename: &str) -> BufWriter<Box<dyn Write>> {
-    match settings.filter {
-        None => get_file_writer(filename),
-
-        Some(ref filter_command) => BufWriter::new(Box::new(
-            // spawn a shell command and write to it
-            FilterWriter::new(&filter_command, &filename),
-        ) as Box<dyn Write>),
-    }
-}
-
 fn split(settings: &Settings) -> i32 {
     let mut reader = BufReader::new(if settings.input == "-" {
         Box::new(stdin()) as Box<dyn Read>
@@ -496,7 +357,7 @@ fn split(settings: &Settings) -> i32 {
 
             crash_if_err!(1, writer.flush());
             fileno += 1;
-            writer = instantiate_current_writer(settings, filename.as_str());
+            writer = platform::instantiate_current_writer(&settings.filter, filename.as_str());
             control.request_new_file = false;
             if settings.verbose {
                 println!("creating file '{}'", filename);
